# Implement z-scoring and gap optimization in CE Align