### Acknowledgements

<!--- Please read each of the following items and confirm by replacing
 !--the [ ] with a [X] --->

- [X] I hereby agree to dual licence this and any previous contributions under both
the _Biopython License Agreement_ **AND** the _BSD 3-Clause License_.

- [X] I have read the ``CONTRIBUTING.rst`` file, have run ``pre-commit``
locally, and understand that continuous integration checks will be used to
confirm the Biopython unit tests and style checks pass with these changes.

- [X] I have added my name to the alphabetical contributors listings in the files
``NEWS.rst`` and ``CONTRIB.rst`` as part of this pull request, am listed
already, or do not wish to be listed. (*This acknowledgement is optional.*)

<!--- Briefly describe the changes included in this pull request below
 !--- starting with 'Closes #...' if appropriate --->

### Description

In this pull request, I implement z-score calculation and gap optimization in the PDB module's implementation of CE Align. The **z-score** is a measure of statistical significance or how likely a particular alignment would be encountered by chance. **Gap optimization** is a procedure that iteratively adjusts gaps in an alignment to improve the alignment's RMSD. In the CE Align algorithm, gap optimization is performed when the z-score is above a certain threshold (normally 3.5).

These changes improve the alignment by searching through alternative alignments when the alignment is statistically significant.

### Testing

I adjusted the unit tests for the improved RMSD, and I performed some manual testing.

I used the [MALIDUP](http://prodata.swmed.edu/malidup/) and [MALISAM](http://prodata.swmed.edu/malisam/) databases to manually test these changes.

Here is my testing script:
```python
import os
import time
from itertools import islice
import re
from Bio.PDB import PDBParser, CEAligner
from collections import deque
import pandas as pd


pattern = re.compile("^[a-z0-9]{4}_\\w\\d.pdb$")
parser = PDBParser()
aligner = CEAligner()
results = deque()

for dirpath, _, filenames in islice(os.walk("data/dup"), 1, None):
    dirname = os.path.split(dirpath)[-1]
    pdb_code = dirname[1:5]
    pdb_filenames = list(filename for filename in filenames if pattern.match(filename))

    assert len(pdb_filenames) == 2

    structure_1 = parser.get_structure(pdb_code, os.path.join(dirpath, pdb_filenames[0]))
    structure_2 = parser.get_structure(pdb_code, os.path.join(dirpath, pdb_filenames[1]))

    start_time = time.time()
    aligner.set_reference(structure_1)
    aligner.align(structure_2, transform=False)
    results.append({
        "PDB Code": pdb_code,
        "RMSD": aligner.rms,
        "Length": aligner.length,
        "Time": time.time() - start_time,
        "Did optimize?": aligner.did_optimize,
        "Z Score": aligner.z_score,
    })

results = pd.DataFrame(results)
print(results[["RMSD", "Length", "Time", "Did optimize?", "Z Score"]].describe())
print(results["Did optimize?"].describe())

pattern = re.compile("^[a-z0-9]{4}_\\w{2}.pdb$")
results = deque()

for dirpath, _, filenames in islice(os.walk("data/analogs"), 1, None):
    dirname = os.path.split(dirpath)[-1]
    pdb_filenames = list(filename for filename in filenames if pattern.match(filename))

    assert len(pdb_filenames) == 2

    structure_1 = parser.get_structure(pdb_filenames[0][:4], os.path.join(dirpath, pdb_filenames[0]))
    structure_2 = parser.get_structure(pdb_filenames[1][:4], os.path.join(dirpath, pdb_filenames[1]))

    start_time = time.time()
    aligner.set_reference(structure_1)
    aligner.align(structure_2, transform=False)
    results.append({
        "Directory": dirname,
        "RMSD": aligner.rms,
        "Length": aligner.length,
        "Time": time.time() - start_time,
        "Did optimize?": aligner.did_optimize,
        "Z Score": aligner.z_score,
    })

results = pd.DataFrame(results)
print(results[["RMSD", "Length", "Time", "Did optimize?", "Z Score"]].describe())
print(results["Did optimize?"].describe())
```

### Results with New Code

#### MALIDUP

```
             RMSD      Length        Time     Z Score
count  241.000000  241.000000  241.000000  241.000000
mean     3.287257   83.352697    0.024016    6.248091
std      1.269601   41.113411    0.025839    0.665213
min      0.628612   24.000000    0.001868    4.580000
25%      2.361698   56.000000    0.007987    5.860000
50%      3.189827   72.000000    0.016907    6.230000
75%      4.294839  104.000000    0.027342    6.700000
max      7.228586  256.000000    0.178572    7.930000
count      241
unique       1
top       True
freq       241
Name: Did optimize?, dtype: object
```

#### MALISAM

```
             RMSD      Length        Time    Z Score
count  130.000000  130.000000  130.000000  130.00000
mean     3.764365   61.292308    0.011495    6.69200
std      0.718092   12.932410    0.004768    0.40069
min      2.093043   24.000000    0.003485    5.60000
25%      3.212446   56.000000    0.008451    6.47000
50%      3.726223   56.000000    0.010203    6.70000
75%      4.314141   72.000000    0.013439    6.89250
max      5.365077  112.000000    0.033091    7.64000
count      130
unique       1
top       True
freq       130
Name: Did optimize?, dtype: object
```

### Results with Original Code

#### MALIDUP

```
             RMSD      Length        Time
count  241.000000  241.000000  241.000000
mean     3.449998   83.352697    0.008540
std      1.290120   41.113411    0.011653
min      0.628612   24.000000    0.000344
25%      2.570221   56.000000    0.001848
50%      3.416172   72.000000    0.004262
75%      4.406737  104.000000    0.010367
max      7.300840  256.000000    0.081144
```

#### MALISAM

```
             RMSD      Length        Time
count  130.000000  130.000000  130.000000
mean     3.946192   61.292308    0.004038
std      0.699545   12.932410    0.002931
min      2.480055   24.000000    0.000751
25%      3.409361   56.000000    0.002134
50%      3.872238   56.000000    0.003014
75%      4.476527   72.000000    0.005288
max      5.590935  112.000000    0.015195
```

### Discussion

The calculated z-score is above the threshold (3.5) and gap optimization is performed for all of the alignments in testing. This is unsurprising because the test alignments come from curated databases of homologous domains (MALIDUP) or structurally analogous motifs (MALISAM).

The z-score calculation and the gap optimization increase the total alignment time. However, I did not notice the difference in execution time without timing the test program. In homology searches, gap optimization should not be performed on every alignment because a z-score threshold of 3.5 corresponds to around 1 in 1000 alignments.

As expected, the alignment length is not affected by gap optimization. I included the alignment length as a simple sanity check.

Finally, CE Align with gap optimization does show better RMSDs.

The z-score calculation code is adapted from the [original CE Align code](https://github.com/kad-ecoli/CE). The gap optimization code is written by myself based on the description in the paper. I found it challenging to understand the gap optimization implementation in the original CE Align code.

I implement the z-score calculation in the C extension module because an alignment path's similarity metric is used to calculate the alignment path's z-score. The similarity metric is readily available in the extension module, and the original code was already available in C.

I implement gap optimization in Python because gap optimization is dependent on RMSD calculation, which is readily available in Python and not available in C.

### References

- [CE Align paper](https://pubmed.ncbi.nlm.nih.gov/9796821/)
- [Original CE Align code](https://github.com/kad-ecoli/CE)