@Will-Tyler 2024-06-25
I think I addressed your review comments now. The alignment lengths are the same in the benchmark testing because the gap optimization does not change the length—I just included the length as a sanity check.

Since I made a lot of changes, I reran the test script and the results are still the same.