@Will-Tyler 2024-07-02
I agree with you, and I prefer avoiding introducing state variables. I also don't like creating ad-hoc return types. In my first version of this pull request, I tried to avoid both of these by leaving the optimize code in the `align` function, but long functions are also a problem. I don't know of a quick fix for this, and I'm okay with it as it is.